requirements.txt
sqlalchemy
psycopg2
sqlalchemy_utils
flask
alembic
voluptuous
flask_sqlalchemy
bcrypt
pytest-flask
tokenlib
lazy
